pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ffmpeg-ai"
version = "0.1.0"
description = "AI-powered YouTube Shorts generator with rich CLI"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "httpx>=0.27",
    "openai>=1.30",          # pointed at OpenRouter base URL
    "edge-tts>=6.1",
    "python-dotenv>=1.0",
    "pillow>=10.0",           # image handling before ffmpeg ingestion
    "faster-whisper>=1.0",    # local transcription for captions
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio"]

[project.scripts]
ffmpeg-ai = "ffmpeg_ai.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ffmpeg_ai"]

[tool.ruff]
line-length = 100
target-version = "py311"
